Update ProtectedRoute, cartSlice, Header, firebase.config, Home, user.api, Profile, ForgetPassword, and order.api
